fixing code

+++ 06/PreConcurrent/ExampleRace.cpp
@@ -2,17 +2,17 @@
 
 //"Тяжелый" метод будет по инкриминтировать переменную по адресу. Таким образом для
 //имитации гонки два потока будут обращаться к одной и той же переменной.
-void ExampleRace::DoWork(uint32_t* num, bool mutexOn, int numIterat)
+void ExampleRace::DoWork(uint64_t* num, bool mutexOn, int numIterat)
 {
     if(mutexOn){
-        for(uint32_t i = 0; i < numIterat; i++){
+        for(uint64_t i = 0; i < numIterat; i++){
             mut->lock();
             (*num)++;
             mut->unlock();
         }
     }
     else{
-        for(uint32_t i = 0; i < numIterat; i++){
+        for(uint64_t i = 0; i < numIterat; i++){
             (*num)++;
         }
     }
